refactor: remove unused code and optimize embed message construction
